common: implementing the timeout recovery stratefy
This command is implementing an idea that it is well tested
in plugin like [1] and running in production for years now.

In the years no better solution are been found, so we keep
our incremental timeout for explora in case of failure.

Only explora can change the answer after a while
and the others from our tests are pretty stable (except nakamoto).

In summary if an request fails the retry stratefy try 4 time with an
incremental time, in this way also if the explora will ban us
there is no problem because we have time to recover from the ban.

[1] https://github.com/clightning4j/btcli4j

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>