core: support the recent way to estimate the fees
This add a recent update of core lightning for the estimation fees.

The current way to estimate the fee for core lightning is

Polled by lightningd to get the current feerate, all values must be passed in sat/kVB.

The plugin must return feerate_floor (e.g. 1000 if mempool is empty),
and an array of 0 or more feerates. Each element of feerates is an
object with blocks and feerate, in ascending-blocks order, for example:

```
{
	"feerate_floor": <sat per kVB>,
	"feerates": {
		{ "blocks": 2, "feerate": <sat per kVB> },
		{ "blocks": 6, "feerate": <sat per kVB> },
		{ "blocks": 12, "feerate": <sat per kVB> }
		{ "blocks": 100, "feerate": <sat per kVB> }
	}
}
```

Link: https://github.com/coffee-tools/folgore/issues/26
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>