fix(esplora): change the api for esplora
This commit introduce a simple esplora API that remove the
dependencies over rust-bitcoin and this means that we can
now support liquid.

In addition, support a library that force us to use a specific
types is not a good idea, and due that rust bitcoin is starting
to be a little bit hard to mantain we switch over our implementation.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>