esplora: fix the send raw tx
There are condition like sending a raw tx over and over again
is not always what we want. e.g: if there is an error
we want make sure that the tx is not propagate over and over again.

So this code will help us to make folgore a little bit safier.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>